rag_service/requirements.txt
langchain_community
langchain-openai
langchainhub
langchain
tiktoken
python-dotenv
beautifulsoup4
pymilvus
langchain_milvus
sqlalchemy
fastapi
sqlalchemy-utils
psycopg2
pypdf
langchain-chroma
chromadb 